perf: add initMappingFiles into utils
