+++ scripts/utils.js
@@ -4,6 +4,14 @@ const fs = require('fs-extra');
 
 const handleWithPrefix = (...args) => path.join(__dirname, '../', ...args);
 
+const initMappingFiles = () => {
+  const mappingFilePath = handleWithPrefix('src/assets/mapping.json');
+  const mappingFile = fs.readJsonSync(mappingFilePath).filter(item => item.visible !== false);
+  const assetsFiles = fs.readdirSync(handleWithPrefix('src/assets'));
+  const targetFiles = mappingFile.filter(item => !assetsFiles.includes(item.id));
+  return targetFiles;
+};
+
 const compressJSON = () => {
   glob('dist/**/*.json', (error, files) => {
     if (error) {
@@ -22,4 +30,4 @@ const compressJSON = () => {
   });
 };
 
-module.exports = { handleWithPrefix, compressJSON };
+module.exports = { handleWithPrefix, compressJSON, initMappingFiles };
